+++ backend/pkg/api/handle_kafka_connect.go
@@ -18,7 +18,6 @@ import (
 
 	"github.com/cloudhut/common/rest"
 	con "github.com/cloudhut/connect-client"
-	"github.com/go-chi/chi/v5"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 
@@ -73,7 +72,7 @@ func (api *API) handleGetConnectors() http.HandlerFunc {
 
 func (api *API) handleGetClusterConnectors() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
+		clusterName := rest.GetURLParam(r, "clusterName")
 
 		ctx, cancel := context.WithTimeout(r.Context(), api.ConnectSvc.Cfg.RequestTimeout)
 		defer cancel()
@@ -106,7 +105,7 @@ func (api *API) handleGetClusterConnectors() http.HandlerFunc {
 
 func (api *API) handleGetClusterInfo() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
+		clusterName := rest.GetURLParam(r, "clusterName")
 
 		canSee, restErr := api.Hooks.Authorization.CanViewConnectCluster(r.Context(), clusterName)
 		if restErr != nil {
@@ -139,8 +138,8 @@ func (api *API) handleGetClusterInfo() http.HandlerFunc {
 
 func (api *API) handleGetConnector() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
-		connector := chi.URLParam(r, "connector")
+		clusterName := rest.GetURLParam(r, "clusterName")
+		connector := rest.GetURLParam(r, "connector")
 
 		ctx, cancel := context.WithTimeout(r.Context(), api.ConnectSvc.Cfg.RequestTimeout)
 		defer cancel()
@@ -190,8 +189,8 @@ func (c *putConnectorConfigRequest) ToClientRequest() con.PutConnectorConfigOpti
 
 func (api *API) handlePutConnectorConfig() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
-		connectorName := chi.URLParam(r, "connector")
+		clusterName := rest.GetURLParam(r, "clusterName")
+		connectorName := rest.GetURLParam(r, "connector")
 
 		canEdit, restErr := api.Hooks.Authorization.CanEditConnectCluster(r.Context(), clusterName)
 		if restErr != nil {
@@ -227,8 +226,8 @@ func (api *API) handlePutConnectorConfig() http.HandlerFunc {
 
 func (api *API) handlePutValidateConnectorConfig() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
-		pluginClassName := chi.URLParam(r, "pluginClassName")
+		clusterName := rest.GetURLParam(r, "clusterName")
+		pluginClassName := rest.GetURLParam(r, "pluginClassName")
 
 		canEdit, restErr := api.Hooks.Authorization.CanEditConnectCluster(r.Context(), clusterName)
 		if restErr != nil {
@@ -281,7 +280,7 @@ func (c *createConnectorRequest) ToClientRequest() con.CreateConnectorRequest {
 
 func (api *API) handleCreateConnector() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
+		clusterName := rest.GetURLParam(r, "clusterName")
 
 		canEdit, restErr := api.Hooks.Authorization.CanEditConnectCluster(r.Context(), clusterName)
 		if restErr != nil {
@@ -318,8 +317,8 @@ func (api *API) handleCreateConnector() http.HandlerFunc {
 
 func (api *API) handleDeleteConnector() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
-		connector := chi.URLParam(r, "connector")
+		clusterName := rest.GetURLParam(r, "clusterName")
+		connector := rest.GetURLParam(r, "connector")
 
 		ctx, cancel := context.WithTimeout(r.Context(), api.ConnectSvc.Cfg.RequestTimeout)
 		defer cancel()
@@ -352,8 +351,8 @@ func (api *API) handleDeleteConnector() http.HandlerFunc {
 
 func (api *API) handlePauseConnector() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
-		connector := chi.URLParam(r, "connector")
+		clusterName := rest.GetURLParam(r, "clusterName")
+		connector := rest.GetURLParam(r, "connector")
 
 		ctx, cancel := context.WithTimeout(r.Context(), api.ConnectSvc.Cfg.RequestTimeout)
 		defer cancel()
@@ -386,8 +385,8 @@ func (api *API) handlePauseConnector() http.HandlerFunc {
 
 func (api *API) handleResumeConnector() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
-		connector := chi.URLParam(r, "connector")
+		clusterName := rest.GetURLParam(r, "clusterName")
+		connector := rest.GetURLParam(r, "connector")
 
 		ctx, cancel := context.WithTimeout(r.Context(), api.ConnectSvc.Cfg.RequestTimeout)
 		defer cancel()
@@ -420,8 +419,8 @@ func (api *API) handleResumeConnector() http.HandlerFunc {
 
 func (api *API) handleRestartConnector() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
-		connector := chi.URLParam(r, "connector")
+		clusterName := rest.GetURLParam(r, "clusterName")
+		connector := rest.GetURLParam(r, "connector")
 
 		ctx, cancel := context.WithTimeout(r.Context(), api.ConnectSvc.Cfg.RequestTimeout)
 		defer cancel()
@@ -454,9 +453,9 @@ func (api *API) handleRestartConnector() http.HandlerFunc {
 
 func (api *API) handleRestartConnectorTask() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		clusterName := chi.URLParam(r, "clusterName")
-		connector := chi.URLParam(r, "connector")
-		taskIDstr := chi.URLParam(r, "taskID")
+		clusterName := rest.GetURLParam(r, "clusterName")
+		connector := rest.GetURLParam(r, "connector")
+		taskIDstr := rest.GetURLParam(r, "taskID")
 		taskID, err := strconv.Atoi(taskIDstr)
 		if err != nil {
 			rest.SendRESTError(w, r, api.Logger, &rest.Error{